// __/simple-modal.types.ts
export enum T_MODAL_POSITION {
	LEFT = 'left',
	RIGHT = 'right',
	TOP = 'top',
	BOTTOM = 'bottom',
	CENTER = 'center'
}

export enum T_DEFAULT_VALUES {
	ALL_OF_THEM = 'ALL_OF_THEM',
	OVERRIDE = 'OVERRIDE',
}

export enum T_DEVICE_TYPE {
	MOBILE = 'mobile',
	TABLET = 'tablet',
	DESKTOP = 'desktop',
	CENTER = 'center'
}